Revert "add register useless response to multiple tasks (#5220)"
This reverts commit 64efb668c7a4ebf721f7fedd196f98bd0ec0df7a.
