+++ ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/manager/EthPeers.java
@@ -179,7 +179,7 @@ public class EthPeers {
 
   public void dispatchMessage(
       final EthPeer peer, final EthMessage ethMessage, final String protocolName) {
-    final Optional<RequestManager> maybeRequestManager = peer.dispatch(ethMessage, protocolName);
+    Optional<RequestManager> maybeRequestManager = peer.dispatch(ethMessage, protocolName);
     if (maybeRequestManager.isPresent() && peer.hasAvailableRequestCapacity()) {
       reattemptPendingPeerRequests();
     }
@@ -229,7 +229,7 @@ public class EthPeers {
 
   private void removeDisconnectedPeers() {
     final Collection<EthPeer> peerStream = connections.values();
-    for (final EthPeer p : peerStream) {
+    for (EthPeer p : peerStream) {
       if (p.isDisconnected()) {
         connections.remove(p.getConnection());
       }
