+++ ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/manager/snap/RetryingGetBytecodeFromPeerTask.java
@@ -17,7 +17,6 @@ package org.hyperledger.besu.ethereum.eth.manager.snap;
 import org.hyperledger.besu.ethereum.core.BlockHeader;
 import org.hyperledger.besu.ethereum.eth.manager.EthContext;
 import org.hyperledger.besu.ethereum.eth.manager.EthPeer;
-import org.hyperledger.besu.ethereum.eth.manager.exceptions.IncompleteResultsException;
 import org.hyperledger.besu.ethereum.eth.manager.task.AbstractRetryingPeerTask;
 import org.hyperledger.besu.ethereum.eth.manager.task.EthTask;
 import org.hyperledger.besu.plugin.services.MetricsSystem;
@@ -66,13 +65,7 @@ public class RetryingGetBytecodeFromPeerTask extends AbstractRetryingPeerTask<Ma
     return executeSubTask(task::run)
         .thenApply(
             peerResult -> {
-              if (isEmptyResponse(peerResult.getResult())) {
-                final EthPeer peer = peerResult.getPeer();
-                peer.recordUselessResponse("GetBytecodeFromPeerTask");
-                throw new IncompleteResultsException("No code returned by peer " + peer.nodeId());
-              }
               result.complete(peerResult.getResult());
-              peerResult.getPeer().chainState().updateHeightEstimate(blockHeader.getNumber());
               return peerResult.getResult();
             });
   }
