+++ ethereum/eth/src/test/java/org/hyperledger/besu/ethereum/eth/manager/EthProtocolManagerTestUtil.java
@@ -28,7 +28,6 @@ import org.hyperledger.besu.ethereum.core.ProtocolScheduleFixture;
 import org.hyperledger.besu.ethereum.eth.EthProtocol;
 import org.hyperledger.besu.ethereum.eth.EthProtocolConfiguration;
 import org.hyperledger.besu.ethereum.eth.manager.DeterministicEthScheduler.TimeoutPolicy;
-import org.hyperledger.besu.ethereum.eth.manager.snap.SnapProtocolManager;
 import org.hyperledger.besu.ethereum.eth.peervalidation.PeerValidator;
 import org.hyperledger.besu.ethereum.eth.sync.SynchronizerConfiguration;
 import org.hyperledger.besu.ethereum.eth.transactions.TransactionPool;
@@ -72,7 +71,7 @@ public class EthProtocolManagerTestUtil {
       final EthProtocolConfiguration ethereumWireProtocolConfiguration,
       final Optional<MergePeerFilter> mergePeerFilter) {
 
-    final EthPeers peers =
+    EthPeers peers =
         new EthPeers(
             EthProtocol.NAME,
             () -> protocolSchedule.getByBlockHeader(blockchain.getChainHeadHeader()),
@@ -80,9 +79,9 @@ public class EthProtocolManagerTestUtil {
             new NoOpMetricsSystem(),
             25,
             EthProtocolConfiguration.DEFAULT_MAX_MESSAGE_SIZE);
-    final EthMessages messages = new EthMessages();
-    final EthScheduler ethScheduler = new DeterministicEthScheduler(TimeoutPolicy.NEVER_TIMEOUT);
-    final EthContext ethContext = new EthContext(peers, messages, ethScheduler);
+    EthMessages messages = new EthMessages();
+    EthScheduler ethScheduler = new DeterministicEthScheduler(TimeoutPolicy.NEVER_TIMEOUT);
+    EthContext ethContext = new EthContext(peers, messages, ethScheduler);
 
     return new EthProtocolManager(
         blockchain,
@@ -186,7 +185,7 @@ public class EthProtocolManagerTestUtil {
       final WorldStateArchive worldStateArchive,
       final TransactionPool transactionPool,
       final EthProtocolConfiguration configuration) {
-    final EthPeers peers =
+    EthPeers peers =
         new EthPeers(
             EthProtocol.NAME,
             () -> protocolSchedule.getByBlockHeader(blockchain.getChainHeadHeader()),
@@ -194,7 +193,7 @@ public class EthProtocolManagerTestUtil {
             new NoOpMetricsSystem(),
             25,
             EthProtocolConfiguration.DEFAULT_MAX_MESSAGE_SIZE);
-    final EthMessages messages = new EthMessages();
+    EthMessages messages = new EthMessages();
 
     return create(
         blockchain,
@@ -215,7 +214,7 @@ public class EthProtocolManagerTestUtil {
       final TransactionPool transactionPool,
       final EthProtocolConfiguration configuration,
       final ForkIdManager forkIdManager) {
-    final EthPeers peers =
+    EthPeers peers =
         new EthPeers(
             EthProtocol.NAME,
             () -> protocolSchedule.getByBlockHeader(blockchain.getChainHeadHeader()),
@@ -223,7 +222,7 @@ public class EthProtocolManagerTestUtil {
             new NoOpMetricsSystem(),
             25,
             EthProtocolConfiguration.DEFAULT_MAX_MESSAGE_SIZE);
-    final EthMessages messages = new EthMessages();
+    EthMessages messages = new EthMessages();
 
     return create(
         blockchain,
@@ -241,7 +240,7 @@ public class EthProtocolManagerTestUtil {
       final ProtocolSchedule protocolSchedule,
       final Blockchain blockchain,
       final EthScheduler ethScheduler) {
-    final EthPeers peers =
+    EthPeers peers =
         new EthPeers(
             EthProtocol.NAME,
             () -> protocolSchedule.getByBlockHeader(blockchain.getChainHeadHeader()),
@@ -249,7 +248,7 @@ public class EthProtocolManagerTestUtil {
             new NoOpMetricsSystem(),
             25,
             EthProtocolConfiguration.DEFAULT_MAX_MESSAGE_SIZE);
-    final EthMessages messages = new EthMessages();
+    EthMessages messages = new EthMessages();
 
     return create(
         blockchain,
@@ -390,17 +389,6 @@ public class EthProtocolManagerTestUtil {
     return RespondingEthPeer.builder()
         .ethProtocolManager(ethProtocolManager)
         .estimatedHeight(estimatedHeight)
-        .build();
-  }
-
-  public static RespondingEthPeer createPeer(
-      final EthProtocolManager ethProtocolManager,
-      final SnapProtocolManager snapProtocolManager,
-      final long estimatedHeight) {
-    return RespondingEthPeer.builder()
-        .ethProtocolManager(ethProtocolManager)
-        .estimatedHeight(estimatedHeight)
-        .snapProtocolManager(snapProtocolManager)
         .build();
   }
